+++ src/node/mempool_persist.h
@@ -2,15 +2,15 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#ifndef BITCOIN_KERNEL_MEMPOOL_PERSIST_H
-#define BITCOIN_KERNEL_MEMPOOL_PERSIST_H
+#ifndef BITCOIN_NODE_MEMPOOL_PERSIST_H
+#define BITCOIN_NODE_MEMPOOL_PERSIST_H
 
 #include <util/fs.h>
 
 class Chainstate;
 class CTxMemPool;
 
-namespace kernel {
+namespace node {
 
 /** Dump the mempool to a file. */
 bool DumpMempool(const CTxMemPool& pool, const fs::path& dump_path,
@@ -28,7 +28,7 @@ bool LoadMempool(CTxMemPool& pool, const fs::path& load_path,
                  Chainstate& active_chainstate,
                  ImportMempoolOptions&& opts);
 
-} // namespace kernel
+} // namespace node
 
 
-#endif // BITCOIN_KERNEL_MEMPOOL_PERSIST_H
+#endif // BITCOIN_NODE_MEMPOOL_PERSIST_H
