+++ src/test/fuzz/validation_load_mempool.cpp
@@ -2,7 +2,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include <kernel/mempool_persist.h>
+#include <node/mempool_persist.h>
 
 #include <node/mempool_args.h>
 #include <node/mempool_persist_args.h>
@@ -21,8 +21,8 @@
 #include <cstdint>
 #include <vector>
 
-using kernel::DumpMempool;
-using kernel::LoadMempool;
+using node::DumpMempool;
+using node::LoadMempool;
 
 using node::MempoolPath;
